requirements.txt
pytest==6.2.5
selenium==4.1.0
webdriver-manager==3.8.1
google-api-python-client==2.55.0
retrying==1.3.3
Faker==10.0.0